PXE server installation playbook (#9)
* PXE server installation playbook

Introduce playbooks meant to assist an operator to get started with
Tateru by installing the PXE server with all the needed data.

See https://github.com/tateru/tateru-installer/pull/6 for longer
reasoning why the copytoram size is specified.

* Add Tateru boot service readme

* Decrease loglevel to errors by default

Docker emits an overlayfs warning on boot, this makes it quiet.
It seems reasonable anyhow, this is not the production kernel the users
will end up running in production so it does not matter.

* Introduce TATERU_BOOT_URL

This variable controlls how the URL is constructed for loading the
installer. It allows setting to http/https and either IP addresses or
even DNS if those are configured.